Search function done, fixed table layouts

+++ client/src/components/Dashboard/Products/Products.js
@@ -16,6 +16,7 @@ export default function Products(props) {
   const [newProductPrice, setNewProductPrice] = useState("0.00");
   const [newProductQuantity, setNewProductQuantity] = useState("0.0");
   const [customError, setCustomError] = useState();
+  const [searchTerm, setSearchTerm] = useState("");
   const route = "/products";
   const title = "Products";
   const getInitialDB = async () => {
@@ -235,7 +236,16 @@ export default function Products(props) {
         </div>
         <div className="searchbar">
           <label htmlFor="searchBar"></label>
-          <input type="text" name="searchBar" id="searchBar" placeholder={`Search ${title}`}></input>
+          <input
+            type="text"
+            name="searchBar"
+            id="searchBar"
+            placeholder={`Search ${title}`}
+            onChange={(e) => {
+              setSearchTerm(e.target.value);
+            }}
+            value={searchTerm}
+          ></input>
         </div>
       </div>
       {loading ? null : (
@@ -251,32 +261,36 @@ export default function Products(props) {
             </thead>
             <tbody>
               {products.map((product) => {
-                return (
-                  <tr key={uniqid()}>
-                    <td>{product.name}</td>
-                    <td>{product.quantity.toFixed(1)}</td>
-                    <td>{"$" + product.price.toFixed(2)}</td>
-                    <td>
-                      <div className="actionButtonContainer">
-                        <FaRegEdit
-                          className="actionButton"
-                          size="18"
-                          onClick={() => {
-                            handleEdit(product);
-                            setmenuStateCreate(false);
-                          }}
-                        ></FaRegEdit>
-                        <FaRegTrashAlt
-                          className="actionButton"
-                          size="18"
-                          onClick={() => {
-                            handleDelete(product._id);
-                          }}
-                        ></FaRegTrashAlt>
-                      </div>
-                    </td>
-                  </tr>
-                );
+                if (searchTerm === "" || product.name.toLowerCase().includes(searchTerm.toLowerCase())) {
+                  return (
+                    <tr key={uniqid()}>
+                      <td>{product.name}</td>
+                      <td>{product.quantity.toFixed(1)}</td>
+                      <td>{"$" + product.price.toFixed(2)}</td>
+                      <td>
+                        <div className="actionButtonContainer">
+                          <FaRegEdit
+                            className="actionButton"
+                            size="18"
+                            onClick={() => {
+                              handleEdit(product);
+                              setmenuStateCreate(false);
+                            }}
+                          ></FaRegEdit>
+                          <FaRegTrashAlt
+                            className="actionButton"
+                            size="18"
+                            onClick={() => {
+                              handleDelete(product._id);
+                            }}
+                          ></FaRegTrashAlt>
+                        </div>
+                      </td>
+                    </tr>
+                  );
+                } else {
+                  return null;
+                }
               })}
             </tbody>
           </table>
